fix: rename packae name
